requirements.txt
attr==0.3.2
attrs==21.4.0
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.3.0
cryptography==3.4.8
Cython==0.29.28
dl==0.1.0
docutils==0.17.1
HTMLParser==0.0.2
importlib_metadata==4.11.3
ipython==8.12.0
ipywidgets==7.6.5
Jinja2==2.11.3
jnius==1.1.0
keyring==23.4.0
kornia==0.6.10
lockfile==0.12.2
lxml==4.8.0
matplotlib==3.5.1
numpy==1.21.5
ordereddict==1.1
Pillow==9.4.0
protobuf==4.22.1
pyOpenSSL==23.1.1
railroad==0.5.0
Sphinx==4.4.0
toml==0.10.2
torch==2.0.0
torchvision==0.15.0
tornado==6.1
tqdm==4.64.0
xmlrpclib==1.0.1
zipp==3.7.0
